(#6, #7) Finished CommandPools
We added the CommandBuffer class, constructible from a CommandPool, which finishes that part of the code. We also adapted Buffer::copy() to take advantange of this. The only things that remains is an Array-has-no-copy-constructor bug when returning an Array using a non-copy-constructible class (like a CommandBuffer), but we won't fix it for now as we just commented the multi-buffer allocation function.